+++ src/App.tsx
@@ -93,6 +93,7 @@ const App: React.FC = () => {
     }, [viewerB]);
     const [viewerAReady, setViewerAReady] = useState(false);
     const [viewerBReady, setViewerBReady] = useState(false);
+    const [syncEnabled, setSyncEnabled] = useState(false);
 
     // Generic file input hook.
     function useFileInput<T>(
@@ -135,6 +136,7 @@ const App: React.FC = () => {
     const [chainIdsAlignedTo, setChainIdsAlignedTo] = useState<string[]>([]);
     const [selectedChainIdAlignedTo, setSelectedChainIdAlignedTo] = useState<string>('');
     const [chainIdsAligned, setChainIdsAligned] = useState<string[]>([]);
+    const [selectedChainIdAligned, setSelectedChainIdAligned] = useState<string>('');
 
     // Handle file changes for molecule loading.
     type FileChangeMode = 'alignedTo' | 'aligned';
@@ -610,87 +612,110 @@ const App: React.FC = () => {
         }
     }, [colorsAlignedFile.data, viewerA.moleculeAligned, viewerB.moleculeAligned]);
 
-    // Chain ID selection state.
+    // Effect for moleculeAlignedTo Chain ID selection.
     useEffect(() => {
         console.log('Updating chain IDs for moleculeAlignedTo');
         const pluginA = viewerA.ref.current;
         if (!pluginA) return;
-        const structuresA = pluginA.managers.structure.hierarchy.current.structures;
-        if (!structuresA) {
-            console.warn('No structures found in viewer A.');
-            return;
-        }
-        console.log(`Structures in viewer A:`, structuresA);
-        console.log(`Number of structures:` + structuresA.length);
-        const structureAAlignedTo = structuresA[0].cell.obj?.data;
+        const structure = pluginA.managers.structure.hierarchy.current.structures[0].cell.obj?.data;
         setChainIdsAlignedTo(
-            getChainIdsFromStructure(structureAAlignedTo!)
+            getChainIdsFromStructure(structure!)
         );
     }, [viewerA.moleculeAlignedTo]);
 
-       
+    // Effect for moleculeAligned Chain ID selection.
+    useEffect(() => {
+        console.log('Updating chain IDs for moleculeAligned');
+        const pluginB = viewerB.ref.current;
+        if (!pluginB) return;
+        const structure = pluginB.managers.structure.hierarchy.current.structures[1].cell.obj?.data;
+        setChainIdsAligned(
+            getChainIdsFromStructure(structure!)
+        );
+    }, [viewerB.moleculeAligned]);
+
     /**
-     * Creates a handler to select and zoom to a specific region in both viewers.
+     * Creates a handler to zoom to a chain.
+     * @param structureIndex The index of the structure in the viewer hierarchy to zoom to.
      * @return An object with a handleButtonClick method.
      */
-    function createSelectAndZoom() {
+    function createZoomA(structureIndex: number, chain: string) {
         return {
             handleButtonClick: async () => {
                 const pluginA = viewerA.ref.current;
                 if (!pluginA) return;
-                const pluginB = viewerB.ref.current;
-                if (!pluginB) return;
-                const chain = selectedChainIdAlignedTo;
-                //const chain = selectedChain;
-                // Get structures from viewer A.
-                const structuresA = pluginA.managers.structure.hierarchy.current.structures;
-                if (!structuresA) {
-                    console.warn('No structures found in viewer A.');
+                const structure = pluginA.managers.structure.hierarchy.current.structures[structureIndex].cell.obj?.data;
+                if (!structure) {
+                    console.warn('No structure data found for alignedTo.');
                     return;
                 }
-                console.log(`Structures in viewer A:`, structuresA);
-                console.log(`Number of structures:` + structuresA.length);
-                const structureAAlignedTo = structuresA[0].cell.obj?.data;
-                if (!structureAAlignedTo) {
-                    console.warn('No structure data found for alignedTo in viewer A.');
-                    return;
-                }
-                // const structureAAligned = structuresA[1].cell.obj?.data;
-                // if (!structureAAligned) {
-                //     console.warn('No structure data found for aligned in viewer A.');
-                //     return;
-                // }
                 const qb = MolScriptBuilder.struct.generator.atomGroups({
                     'chain-test': MolScriptBuilder.core.rel.eq([
                         MolScriptBuilder.struct.atomProperty.macromolecular.auth_asym_id(),
                         chain
                     ])
-                    // For pdb_strand_id, use:
-                    // MolScriptBuilder.struct.atomProperty.macromolecular.pdb_strand_id()
                 });
-                // const qb = MolScriptBuilder.struct.generator.atomGroups({
-                //     'chain-test': MolScriptBuilder.core.rel.eq([
-                //         MolScriptBuilder.struct.atomProperty.macromolecular.auth_asym_id(),
-                //         'A'
-                //     ]),
-                //     'residue-test': MolScriptBuilder.core.rel.inRange([
-                //         MolScriptBuilder.struct.atomProperty.macromolecular.auth_seq_id(),
-                //         10, 20
-                //     ])
-                // });
                 const compiled = compile(qb);
-                const ctx = new QueryContext(structureAAlignedTo);
+                const ctx = new QueryContext(structure);
                 const selection = compiled(ctx);
                 const loci = StructureSelection.toLociWithSourceUnits(selection);
-
+                // Zoom
                 pluginA.managers.camera.focusLoci(loci);
-                pluginB.managers.camera.focusLoci(loci);
+                if (syncEnabled){
+                    const pluginB = viewerB.ref.current;
+                    if (!pluginB) return;
+                    pluginB.managers.camera.focusLoci(loci);
+                }
             }
         };
     }
-    // Create select and zoom handlers.
-    const selectAndZoomA = createSelectAndZoom();
-    const selectAndZoomB = createSelectAndZoom();
+
+    // Create zoomA handlers.
+    const zoomAAlignedTo = createZoomA(0, selectedChainIdAlignedTo);
+    const zoomAAligned = createZoomA(1, selectedChainIdAligned);
+
+    /**
+     * Creates a handler to zoom to a chain.
+     * @return An object with a handleButtonClick method.
+     * 
+     * @param structureIndex The index of the structure in the viewer hierarchy to zoom to.
+     * @return An object with a handleButtonClick method.
+     */
+    function createZoomB(structureIndex: number, chain: string) {
+        return {
+            handleButtonClick: async () => {
+                const pluginB = viewerB.ref.current;
+                if (!pluginB) return;
+                // Get structure.
+                const structure = pluginB.managers.structure.hierarchy.current.structures[structureIndex].cell.obj?.data;
+                if (!structure) {
+                    console.warn('No structure data found for aligned.');
+                    return;
+                }
+                const qb = MolScriptBuilder.struct.generator.atomGroups({
+                    'chain-test': MolScriptBuilder.core.rel.eq([
+                        MolScriptBuilder.struct.atomProperty.macromolecular.auth_asym_id(),
+                        chain
+                    ])
+                });
+                const compiled = compile(qb);
+                const ctx = new QueryContext(structure);
+                const selection = compiled(ctx);
+                const loci = StructureSelection.toLociWithSourceUnits(selection);
+
+                pluginB.managers.camera.focusLoci(loci);
+                if (syncEnabled){
+                    const pluginA = viewerA.ref.current;
+                    if (!pluginA) return;
+                    pluginA.managers.camera.focusLoci(loci);
+                }
+            }
+        };
+    }
+
+    // Create zoomB handlers.
+    const zoomBAlignedTo = createZoomB(0, selectedChainIdAlignedTo);
+    const zoomBAligned = createZoomB(1, selectedChainIdAligned);
     
     // Return the main app component.
     return (
@@ -777,13 +802,22 @@ const App: React.FC = () => {
                         ref={colorsAlignedFile.inputRef}
                         onChange={colorsAlignedFile.handleFileChange}
                     />
+                    <ChainSelectButton
+                        disabled={!viewerB.isMoleculeAlignedLoaded}
+                        chainIds={chainIdsAligned}
+                        selectedChainId={selectedChainIdAligned}
+                        onSelect={setSelectedChainIdAligned}
+                        label="Select Chain"
+                    />
                 </div>
                 <div>
                     <SyncButton
                         viewerA={viewerA.ref.current}
                         viewerB={viewerB.ref.current}
                         activeViewer={activeViewer}
-                        disabled={!viewerB.isMoleculeAlignedLoaded}
+                        disabled={!viewerB.isMoleculeAlignedToLoaded}
+                        syncEnabled={syncEnabled}
+                        setSyncEnabled={setSyncEnabled}
                     />
                     <button
                         onClick={dictionaryFile.handleButtonClick}
@@ -826,6 +860,12 @@ const App: React.FC = () => {
                                 </span>
                             </button>
                             <button
+                                onClick={zoomAAlignedTo.handleButtonClick}
+                                disabled={!selectedChainIdAlignedTo}
+                            >
+                                Zoom to: {selectedChainIdAlignedTo}
+                            </button>
+                            <button
                                 onClick={toggleViewerAAligned.handleButtonClick}
                                 disabled={!viewerA.isMoleculeAlignedLoaded}
                             >
@@ -835,10 +875,10 @@ const App: React.FC = () => {
                                 </span>
                             </button>
                             <button
-                                onClick={selectAndZoomA.handleButtonClick}
-                                disabled={!selectedChainIdAlignedTo}
+                                onClick={zoomAAligned.handleButtonClick}
+                                disabled={!selectedChainIdAligned}
                             >
-                                Zoom to: {selectedChainIdAlignedTo}
+                                Zoom to: {selectedChainIdAligned}
                             </button>
                         </div>
                         <MolstarContainer
@@ -860,6 +900,12 @@ const App: React.FC = () => {
                                 </span>
                             </button>
                             <button
+                                onClick={zoomBAlignedTo.handleButtonClick}
+                                disabled={!selectedChainIdAlignedTo}
+                            >
+                                Zoom to: {selectedChainIdAlignedTo}
+                            </button>
+                            <button
                                 onClick={toggleViewerBAligned.handleButtonClick}
                                 disabled={!viewerB.isMoleculeAlignedLoaded}
                             >
@@ -869,10 +915,10 @@ const App: React.FC = () => {
                                 </span>
                             </button>
                             <button
-                                onClick={selectAndZoomB.handleButtonClick}
-                                disabled={!isMoleculeAlignedColoursLoaded}
+                                onClick={zoomBAligned.handleButtonClick}
+                                disabled={!selectedChainIdAligned}
                             >
-                                Select and Zoom
+                                Zoom to: {selectedChainIdAligned}
                             </button>
                         </div>
                         <MolstarContainer