+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## [v0.4.4] - 2025-12-22
+
+Added `Select Chain` buttons to select chains for AlignedTo and Aligned data.
+The `Select and Zoom` is replaced by `Zoom to:` buttons for each molecule in each viewer.
+If the viewers are set to sync, then the zoom happens in both. Otherwise the zoom happens only in the viewer where the `Zoom to:` button is actioned.
+
 ## [v0.4.3] - 2025-12-20
 
 Updated README.md
 Added LICENSE
-Select and zoom now zoms to a specific chain rather than the first atom.  
+`Select and Zoom` button now zooms to a specific chain rather than the first atom.